Create useSearchForm, refactoring
