+++ client/src/components/SearchForm/SearchForm.jsx
@@ -3,7 +3,6 @@ import Button from 'components/Button/Button'
 import Input from 'components/Input/Input'
 
 export default function SearchForm({
-  onChange,
   onSubmit,
   isDisabled,
   search,
